[TW14850757] Delete `.git/hooks` when composer update/install runs.
TLDR: This is to prevent an issue with the Pattern Lab repository.

---

Courtyard (Pattern Lab) is installed with Composer, which removes any `.git` directories via post install/upate commands.  “Courtyard” uses Husky to run linting via Git hooks, which gets installed when you run `npm install`, and expects that `.git` repo to exist.  When it is not found, Husky looks up directories until it finds one, and installs there.  We need to prevent that, since it messes up code formatting, so these commands delete the hooks entirely.