[TW14850757] Add supported regions, and site logo.
Note: Ideally we would be able to use the site branding block, and potentially even the header component, but that’s not currently possible with the way it’s been written at the moment. I tried to at least use the branding block first, but due to the way the CSS is being written, wrapper markup is not going to work. We are also not actually allowing a user-defined image to be used anyway, so I don’t think it’s worth it.