SRL-369: explicitly set values for node basic text section
This previously used content|without(...) to omit fields it didn't want, but
adding new fields to the content and not remembering to omit it causes
it to appear where it shouldn't in the node.

By explicitly defining fields we want to send to that template we make
the code more maintainable, and prevent this issue from happening again.

Would recommend refactoring this way anywhere this is found.