SRL-445: Minor patch to linkit module
Linkit EntityMatcher used LIKE in a query condition which is not supported.
Patch updates the condition to use CONTAINS.
There was some discussion on the 8.5 fix about config entities vs content
entities and a conditional to check the entity class and use LIKE only for
content entities, but that doesn't seem to be supported. I'm not aware of
any reason not to use CONTAINS on all queries, so I have not ported the
8.5 patch to our 8.4 module.